// lib/api/generated/models/TagApiDeleteMultipleTagRelationValuesRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { TagApiDeleteTagRelationValueItem } from './TagApiDeleteTagRelationValueItem';
export type TagApiDeleteMultipleTagRelationValuesRequest = {
  tagRelationName?: string;
  items?: Array<TagApiDeleteTagRelationValueItem>;
};