// lib/api/generated/models/TagApiDeleteMultipleTagValuesRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { TagApiDeleteTagValueItem } from './TagApiDeleteTagValueItem';
export type TagApiDeleteMultipleTagValuesRequest = {
  tagName?: string;
  items?: Array<TagApiDeleteTagValueItem>;
};