// lib/api/generated/models/TagApiGetMultipleTagRelationValueRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TagApiGetMultipleTagRelationValueRequest = {
  tagRelationNames?: Array<string>;
  startTime?: string;
  endTime?: string;
  query?: string;
  connectionTimeout?: number;
};