// lib/api/generated/models/TagApiGetMultipleTagValueRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TagApiGetMultipleTagValueRequest = {
  tagNames?: Array<string>;
  startTime?: string;
  endTime?: string;
  query?: string;
  connectionTimeout?: number;
};